backend/requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
bleach==6.0.0
cffi==1.15.1
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2
django-filter==23.1
django-summernote==0.8.20.0
django-tinymce==3.6.1
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Markdown==3.4.3
Pillow==9.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyenchant==3.2.2
pyOpenSSL==23.1.1
pytz==2023.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
webencodings==0.5.1
zope.interface==6.0
python-dotenv==1.0.0
django-resized==1.0.2